take 765

+++ Src/Xt.Synth0.DSP/DSP/UnitDSP.cpp
@@ -22,7 +22,7 @@ UnitDSP::Freq(UnitModel const& model, SynthInput const& input)
 	int key = input.oct * 12 + static_cast<int>(input.note);
 	int unit = (model.oct + 1) * 12 + static_cast<int>(model.note);
 	int noteNum = unit + key - base;
-	int cent = static_cast<int>(Mix0100Inclusive(model.dtn));
+	int cent = Mix0100Inclusive(model.dtn);
 	float midi = noteNum + cent / 100.0f;
 	return 440.0f * powf(2.0f, (midi - 69.0f) / 12.0f);
 }
